refactor extraworksheethtml css
