+++ frontend/src/components/worksheets/Worksheet/Worksheet.js
@@ -6,7 +6,7 @@ import { keepPosInView, renderPermissions, getMinMaxKeys } from '../../../util/w
 import * as Mousetrap from '../../../util/ws_mousetrap_fork';
 import WorksheetItemList from '../WorksheetItemList';
 import ReactDOM from 'react-dom';
-import ExtraWorksheetHTML from '../ExtraWorksheetHTML';
+import ExtraWorksheetHTML from '../ExtraWorksheetHTML/ExtraWorksheetHTML';
 import 'bootstrap';
 import 'jquery-ui-bundle';
 import WorksheetHeader from './WorksheetHeader';
