+++ frontend/src/css/imports.scss
@@ -551,27 +551,6 @@ body {
   z-index: 1060;
 }
 
-.progress-message {
-  display: none;
-  position: fixed;
-  bottom: 40px;
-  right: 40px;
-  width: 150px;
-  height: 40px;
-  line-height: 40px;
-  background-color: rgba(255, 255, 255, 0.75);
-  text-align: center;
-  z-index: 1020;
-}
-.progress-message#update_progress {
-  top: 63px;
-}
-.progress-message#save_error i {
-  font-size: 18px;
-  color: red;
-  vertical-align: -3px;
-  margin-right: 2px;
-}
 #worksheet-editor {
   margin: 0;
   position: relative;
@@ -581,17 +560,6 @@ body {
   right: 0;
   width: 1000px;
   height: 700px;
-}
-#glossaryModal .modal-header h4 {
-  margin: 0;
-}
-#glossaryModal h5 {
-  margin-top: 0;
-}
-#glossaryModal .table > tbody > tr > td {
-  padding: 5px 6px 8px 6px;
-  border-color: #eee;
-  border-bottom: 1px solid #eee;
 }
 #bundle-upload-form {
   display: none;
